@@ -12,12 +12,10 @@
 		"@itsfuad/domtoastmessage": "^1.0.7",
 		"@sveltejs/adapter-auto": "^3.0.0",
 		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"@vercel/analytics": "^1.3.1",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"sass": "^1.69.5",
-		"svelte": "^4.2.7",
+		"svelte": "^5",
 		"svelte-check": "^3.6.0",
-		"typescript": "^5.0.0",
 		"vite": "^5.0.3"
 	},
 	"type": "module"
